Add default values for Watch description and image URL
This commit introduces a check to set a default description of "No description provided." if the `Description` field in `viewModel.WatchItems` is empty. It also validates the `ImageUrl`, using a default image URL ("images/no-picture-Square210.png") if the provided URL is invalid. These changes improve the robustness of the `BuildWatchAsync` method.